txt simples criaçao/exclusao do documento

+++ ProjetoModulo05arquivos/Form1.cs
@@ -24,9 +24,40 @@ namespace ProjetoModulo05arquivos
         private void btnCriar_Click(object sender, EventArgs e)
         {
             //variavel para o nome do arquivo que sera criado
-            string nomeArq = "Arquivo.txt";
-            //criaçao do arquivo
-            Stream arquivo = File.Create(nomeArq);
+            //para reconhecer o local deve-se colocar um @ antes ou colocar \\
+            string nomeArq = @"C:\Users\adm\Documents\Visual Studio 2017\exercicios\Curso Forms\ProjetoModulo05arquivos\ProjetoModulo05arquivos\bin\Debug\Arquivos\Arquivo.txt";
+            //se o arquivo ja existir aparece a mensagem se não ele aparece outra mensagem e realiza a criaçao
+            if(File.Exists(nomeArq))
+            {
+                MessageBox.Show("Arquivo existe");
+            }
+            else
+            {
+                MessageBox.Show("Arquivo não existe");
+                //criaçao do arquivo
+                Stream arquivo = File.Create(nomeArq);
+                //fecha o arquivo(para que nao fique em uso)
+                arquivo.Close();
+            }
+           
+            
+        }
+
+        private void btnExcluir_Click(object sender, EventArgs e)
+        {
+            string nomeArq = @"C:\Users\adm\Documents\Visual Studio 2017\exercicios\Curso Forms\ProjetoModulo05arquivos\ProjetoModulo05arquivos\bin\Debug\Arquivos\Arquivo.txt";
+            if (File.Exists(nomeArq))
+            {
+                MessageBox.Show("Arquivo existe");
+                //Exclui o arquivo
+                File.Delete(nomeArq);
+            }
+            else
+            {
+                MessageBox.Show("Arquivo não existe");
+             
+               
+            }
         }
     }
 }
